+++ lib/engine.rb
@@ -1,0 +1,5 @@
+# Defining the engine class.
+class Engine
+  def initilize
+  end
+end